+++ docs/CHANGELOG.md
@@ -1,4 +1,37 @@
-## Version 0.1.4
+---
+sidebar_position: 3
+---
+
+# Changelog
+
+## 0.1.0
+
+- Initial release
+
+
+## 0.1.1
+
+Features:
+- Added support for zsh
+- Added git status module
+
+## 0.1.2
+
+Features:
+- Added os icon module
+- Added $exit_code_color variable
+
+Bug Fixes:
+- Fixed a bug where the exit code was not being set correctly
+
+## 0.1.3
+
+Improvements:
+- Set up CI/CD pipeline
+- Added documentation
+- Added changelog
+
+## 0.1.4
 
 Improvements:
 - Expanded documentation